Refactor combined experiments into its own helper
