+++ lib/split.rb
@@ -13,6 +13,7 @@ require 'split/experiment_catalog'
 require 'split/extensions/string'
 require 'split/goals_collection'
 require 'split/helper'
+require 'split/combined_experiments_helper'
 require 'split/metric'
 require 'split/persistence'
 require 'split/redis_interface'
