+++ lib/split/engine.rb
@@ -5,6 +5,8 @@ module Split
       if Split.configuration.include_rails_helper
         ActionController::Base.send :include, Split::Helper
         ActionController::Base.helper Split::Helper
+        ActionController::Base.send :include, Split::CombinedExperimentsHelper
+        ActionController::Base.helper Split::CombinedExperimentsHelper
       end
     end
   end
